_reference/GraphVisualizer/requirements.txt
# Core dependencies
python-dotenv
openai
pydantic

# Database
sqlalchemy
neo4j

# Web framework
flask
jinja2

# Markdown rendering
markdown2

# Async support
aiohttp

# Vector embeddings
sentence-transformers
numpy

# Development dependencies
pytest
pytest-asyncio
